Changed up display for general info, displays habitat and removing bait and time

+++ app/src/main/java/com/example/parakeet/Parakeet_Database/FishDAO.java
@@ -23,5 +23,5 @@ public interface FishDAO {
     LiveData<List<Fish>> getAllFish();
 
     @Query("SELECT * FROM " + FishDatabase.FISH_TABLE + " WHERE fish_user_id = :userId")
-    LiveData<List<Fish>> getAllFishByUserId(int userId);
+    LiveData<List<Fish>> getAllFishByUserId(long userId);
 }
